使用阿里代码规范。规范代码写法

+++ src/main/java/com/genersoft/iot/vmp/conf/UserSetting.java
@@ -1,5 +1,6 @@
 package com.genersoft.iot.vmp.conf;
 
+import io.swagger.models.auth.In;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
@@ -19,7 +20,7 @@ public class UserSetting {
 
     private Boolean seniorSdp = Boolean.FALSE;
 
-    private Long playTimeout = 18000L;
+    private Integer playTimeout = 18000;
 
     private int platformPlayTimeout = 60000;
 
@@ -55,7 +56,7 @@ public class UserSetting {
         return seniorSdp;
     }
 
-    public Long getPlayTimeout() {
+    public Integer getPlayTimeout() {
         return playTimeout;
     }
 
@@ -83,7 +84,7 @@ public class UserSetting {
         this.seniorSdp = seniorSdp;
     }
 
-    public void setPlayTimeout(Long playTimeout) {
+    public void setPlayTimeout(Integer playTimeout) {
         this.playTimeout = playTimeout;
     }
